requirements.txt
# Telegram
aiogram>=3.4.0,<4.0.0
telethon>=1.34.0,<2.0.0

# Database
sqlalchemy[asyncio]>=2.0.27,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
pgvector>=0.2.5,<1.0.0

# OpenAI
openai>=1.12.0,<2.0.0

# Configuration
pyyaml>=6.0.1,<7.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.4.1,<2.6.0
pydantic-settings>=2.2.0,<3.0.0

# Utils
cryptography>=42.0.0,<43.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
